news_genie/requirements.txt
langchain
langchain-community
langgraph
openai
faiss-cpu
streamlit
python-dotenv
tavily-python